refactor : 공연 문의 컴포넌트 추출
