+++ presentation/src/main/java/com/nexters/boolti/presentation/screen/showdetail/ShowDetailScreen.kt
@@ -18,7 +18,6 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.heightIn
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.wrapContentSize
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.RoundedCornerShape
@@ -26,8 +25,6 @@ import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.Divider
 import androidx.compose.material3.DropdownMenu
 import androidx.compose.material3.DropdownMenuItem
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
@@ -45,12 +42,10 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.toArgb
-import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.platform.LocalClipboardManager
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.stringArrayResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.text.AnnotatedString
 import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.dp
@@ -68,6 +63,7 @@ import com.nexters.boolti.presentation.component.BtAppBar
 import com.nexters.boolti.presentation.component.BtAppBarDefaults
 import com.nexters.boolti.presentation.component.CopyButton
 import com.nexters.boolti.presentation.component.MainButton
+import com.nexters.boolti.presentation.component.ShowInquiry
 import com.nexters.boolti.presentation.extension.requireActivity
 import com.nexters.boolti.presentation.screen.LocalSnackbarController
 import com.nexters.boolti.presentation.screen.ticketing.ChooseTicketBottomSheet
@@ -323,7 +319,6 @@ private fun ContentScaffold(
     modifier: Modifier = Modifier,
 ) {
     val snackbarController = LocalSnackbarController.current
-    var telephoneBottomSheet: Boolean? by remember { mutableStateOf(null) }
 
     Column(
         modifier = modifier,
@@ -409,46 +404,11 @@ private fun ContentScaffold(
         Section(
             title = { SectionTitle(stringResource(id = R.string.ticketing_host)) },
             content = {
-                Row(
-                    horizontalArrangement = Arrangement.spacedBy(20.dp),
-                ) {
-                    Text(
-                        modifier = Modifier.weight(1f),
-                        text = host.substringBefore("("),
-                        style = MaterialTheme.typography.bodyLarge.copy(color = Grey30),
-                    )
-                    IconButton(
-                        modifier = Modifier.size(24.dp),
-                        onClick = { telephoneBottomSheet = true }
-                    ) {
-                        Icon(
-                            modifier = Modifier.size(24.dp),
-                            imageVector = ImageVector.vectorResource(id = R.drawable.ic_telephone),
-                            contentDescription = stringResource(id = R.string.show_call_to_ask),
-                            tint = Grey30
-                        )
-                    }
-                    IconButton(
-                        modifier = Modifier.size(24.dp),
-                        onClick = { telephoneBottomSheet = false }
-                    ) {
-                        Icon(
-                            modifier = Modifier.size(24.dp),
-                            imageVector = ImageVector.vectorResource(id = R.drawable.ic_message),
-                            contentDescription = stringResource(id = R.string.show_text_to_ask),
-                            tint = Grey30
-                        )
-                    }
-                }
+                ShowInquiry(
+                    hostName = host.substringBefore("("),
+                    hostNumber = host.substringAfter("(").substringBefore(")")
+                )
             },
-        )
-    }
-
-    telephoneBottomSheet?.let { isTelephone ->
-        InquiryBottomSheet(
-            isTelephone = isTelephone,
-            onDismissRequest = { telephoneBottomSheet = null },
-            contact = host.substringAfter("(").substringBefore(")")
         )
     }
 }
